// src/app/models/project.model.ts
export interface ProjectImage {
    imageUrl: string;
    orderIndex: number;
    isFeatured: boolean;
}

export interface Project {
    id?: string; // Optional because it's generated by Firestore
    title: string;
    shortDescription: string;
    fullDescription: string;
    featuredImageUrl: string;
    status: 'draft' | 'published';
    showOnHomepage?: boolean;
    createdAt: any;
    updatedAt?: any;
    images?: ProjectImage[];
}